pyproject.toml
[project]
name = "semantic-transmission"
version = "0.1.0"
description = "语义传输预研项目：基于 AI 模型的视频语义级压缩传输"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "websocket-client",
    "Pillow",
    "opencv-python",
    "numpy",
    "transformers>=4.51.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "accelerate>=0.30.0",
    "qwen-vl-utils>=0.0.8",
    "scikit-image>=0.24.0",
    "lpips>=0.1.4",
    "click>=8.0",
    "gradio>=5.0",
    "bitsandbytes>=0.46.1",
    "diffusers>=0.33.0",
    "gguf>=0.10.0",
]

[project.scripts]
semantic-tx = "semantic_transmission.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/semantic_transmission"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
